Renames, tweaks and deletes
