+++ src/resolver-map/utils/highlight-all-callback.ts
@@ -1,0 +1,7 @@
+import { Highlight } from '../../highlight/highlight';
+import { field } from '../../highlight/highlighter/field';
+import { type } from '../../highlight/highlighter/type';
+
+export function highlightAllCallback(h: Highlight): Highlight {
+  return h.include(type(), field());
+}
